+++ NEWS.md
@@ -1,9 +1,11 @@
 
 ## News
 
-### mixedBayes 0.2.0 \[2026-01-01\]
+### mixedBayes 0.2.0 \[2026-01-12\]
 
 - Added function `predict_mixedBayes` for making predictions based on fitted mixedBayes models.
+- Added function `GE` for creating gene-environment interaction terms.
+- Added function `reformat` for creating long format data for longitudinal analysis.
 - Fixed minor bugs.
 - Updated the documentation.